some changes

+++ src/code0599.rs
@@ -48,13 +48,13 @@ impl Solution {
     pub fn find_restaurant(list1: Vec<String>, list2: Vec<String>) -> Vec<String> {
         let mut map = std::collections::HashMap::new();
         for (i, s) in list1.iter().enumerate() {
-            map.insert(s, i as i32);
+            map.insert(s, i);
         }
-        let mut min_sum = std::i32::MAX;
+        let mut min_sum = std::usize::MAX;
         let mut result = Vec::new();
         for (i, s) in list2.iter().enumerate() {
             if let Some(&j) = map.get(s) {
-                let sum = i as i32 + j;
+                let sum = i + j;
                 match sum.cmp(&min_sum) {
                     std::cmp::Ordering::Less => {
                         min_sum = sum;
